Adding the observed and simulated data
Uploading the randomly Simulated data for FRB and observed FRB data by our telescope (observed data have entry and exit time of FRB ). The FRBs are repeated in observed data because we are detecting Repeated FRBs because Repeated FRBs give more information for research.  We are using the Ooty Radio telescope (ORT) (assumed not movable) for this observation.